# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio CMS & ATS CV Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  error?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface VerifyEmailBody {
  token: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  /** @minLength 6 */
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  viewer: "viewer",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  isVerified: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface EducationItem {
  id: string;
  institution: string;
  degree: string;
  field: string;
  startYear: string;
  endYear: string;
  description?: string;
}

export interface ExperienceItem {
  id: string;
  company: string;
  position: string;
  startDate: string;
  endDate?: string;
  isCurrent: boolean;
  description?: string;
  technologies?: string[];
}

export interface SkillItem {
  id: string;
  name: string;
  category: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  level: number;
}

export interface Profile {
  id: string;
  name: string;
  title: string;
  bio: string;
  email?: string;
  phone?: string;
  location?: string;
  avatarUrl?: string;
  githubUrl?: string;
  linkedinUrl?: string;
  websiteUrl?: string;
  education?: EducationItem[];
  experience?: ExperienceItem[];
  skills?: SkillItem[];
  tools?: string[];
  expertiseAreas?: string[];
  typewriterTitles?: string[];
  updatedAt?: string;
}

export interface UpdateProfileBody {
  name?: string;
  title?: string;
  bio?: string;
  email?: string;
  phone?: string;
  location?: string;
  avatarUrl?: string;
  githubUrl?: string;
  linkedinUrl?: string;
  websiteUrl?: string;
  education?: EducationItem[];
  experience?: ExperienceItem[];
  skills?: SkillItem[];
  tools?: string[];
  expertiseAreas?: string[];
  typewriterTitles?: string[];
}

export interface PortfolioProject {
  id: string;
  title: string;
  description: string;
  longDescription?: string;
  category: string;
  technologies: string[];
  imageUrl?: string;
  demoUrl?: string;
  githubUrl?: string;
  featured: boolean;
  order?: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreatePortfolioProjectBody {
  title: string;
  description: string;
  longDescription?: string;
  category: string;
  technologies: string[];
  imageUrl?: string;
  demoUrl?: string;
  githubUrl?: string;
  featured?: boolean;
  order?: number;
}

export type PortfolioStatsByCategoryItem = {
  category: string;
  count: number;
};

export interface PortfolioStats {
  total: number;
  featured: number;
  byCategory: PortfolioStatsByCategoryItem[];
}

export type ContentItemSection =
  (typeof ContentItemSection)[keyof typeof ContentItemSection];

export const ContentItemSection = {
  home: "home",
  about: "about",
  contact: "contact",
  services: "services",
  testimonials: "testimonials",
  custom: "custom",
} as const;

export type ContentItemMetadata = { [key: string]: unknown };

export interface ContentItem {
  id: string;
  title: string;
  body: string;
  section: ContentItemSection;
  published: boolean;
  order?: number;
  metadata?: ContentItemMetadata;
  createdAt?: string;
  updatedAt?: string;
}

export type CreateContentBodySection =
  (typeof CreateContentBodySection)[keyof typeof CreateContentBodySection];

export const CreateContentBodySection = {
  home: "home",
  about: "about",
  contact: "contact",
  services: "services",
  testimonials: "testimonials",
  custom: "custom",
} as const;

export type CreateContentBodyMetadata = { [key: string]: unknown };

export interface CreateContentBody {
  title: string;
  body: string;
  section: CreateContentBodySection;
  published?: boolean;
  order?: number;
  metadata?: CreateContentBodyMetadata;
}

export interface CvPersonal {
  fullName: string;
  email: string;
  phone?: string;
  location?: string;
  linkedin?: string;
  github?: string;
  website?: string;
  summary?: string;
}

export interface CvEducation {
  id: string;
  institution: string;
  degree: string;
  field: string;
  startDate: string;
  endDate?: string;
  gpa?: string;
  description?: string;
}

export interface CvExperience {
  id: string;
  company: string;
  position: string;
  startDate: string;
  endDate?: string;
  isCurrent: boolean;
  bullets: string[];
}

export interface CvSkillGroup {
  id: string;
  category: string;
  skills: string[];
}

export interface CvCertification {
  id: string;
  name: string;
  issuer: string;
  date: string;
  url?: string;
}

export type CvDataLanguagesItem = {
  id: string;
  name: string;
  proficiency: string;
};

export interface CvData {
  id: string;
  label: string;
  personal: CvPersonal;
  education?: CvEducation[];
  experience?: CvExperience[];
  skillGroups?: CvSkillGroup[];
  certifications?: CvCertification[];
  languages?: CvDataLanguagesItem[];
  createdAt?: string;
  updatedAt?: string;
}

export interface BlogPost {
  id: string;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  coverImage?: string;
  tags?: string[];
  published?: boolean;
  publishedAt?: string;
  readingTime?: number;
  views?: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface BlogPostBody {
  title: string;
  excerpt: string;
  content: string;
  coverImage?: string;
  tags?: string[];
  published?: boolean;
}

export interface BlogListResponse {
  posts: BlogPost[];
  total: number;
  page: number;
  pages: number;
}

export interface TrackBody {
  path: string;
  referrer?: string;
  sessionId: string;
}

export interface AnalyticsTopPage {
  path: string;
  count: number;
}

export interface AnalyticsDailyView {
  date: string;
  count: number;
}

export interface AnalyticsTopReferrer {
  referrer: string;
  count: number;
}

export interface AnalyticsSummary {
  totalViews: number;
  views30d: number;
  views7d: number;
  views24h: number;
  uniqueVisitors30d: number;
  topPages: AnalyticsTopPage[];
  dailyViews: AnalyticsDailyView[];
  topReferrers: AnalyticsTopReferrer[];
}

export type CreateCvBodyLanguagesItem = {
  id: string;
  name: string;
  proficiency: string;
};

export interface CreateCvBody {
  label: string;
  personal: CvPersonal;
  education?: CvEducation[];
  experience?: CvExperience[];
  skillGroups?: CvSkillGroup[];
  certifications?: CvCertification[];
  languages?: CreateCvBodyLanguagesItem[];
}

export type ListPortfolioParams = {
  category?: string;
  featured?: boolean;
};

export type ListContentParams = {
  section?: string;
  published?: boolean;
};

export type ListBlogPostsParams = {
  tag?: string;
  search?: string;
  page?: string;
  limit?: string;
};

export type TrackPageView201 = {
  ok: boolean;
};